Module type PatriciaTree.VALUE

Module type used for specifying custom homogeneous value types in MakeCustomMap. For most purposes, use the provided Value implementation. It sets 'a t = 'a, which is the desired effect (maps can map to any value). This is the case in MakeMap. However, for maps like Hash-consed maps and sets, it can be useful to restrict the type of values in order to implement hash and polyeq functions on values. See the HASHED_VALUE module type for more details.

type 'a t

The type of values. A 'map map maps key to 'map value. Can be mutable if desired, unless it is being used in Hash-consed maps and sets.